server/requirements.txt
starlette==0.13.6
pytz==2020.1
PyMySQL==0.10.1
sodapy==2.1.0
pydantic==1.7
fastapi==0.63.0
python-dotenv==0.15.0
python_dateutil==2.8.1
